rename(policy/LottoPolicy): constants/LottoConstants로 이름 변경
해당 이름이 역할에 더 잘 맞는 것 같아 변경하였습니다.
